Move NetworkingExtensions to the SpectatorView assembly
