+++ Assets/MixedRealityToolkit.Extensions/SpectatorView/Scripts/HolographicCamera/CameraPoseProvider.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using Microsoft.MixedReality.Toolkit.Extensions.Experimental.Socketer;
+using Microsoft.MixedReality.Toolkit.Extensions.Experimental.SpectatorView.StateSynchronization;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
