+++ Assets/MixedRealityToolkit.Extensions/SpectatorView/Scripts/StateSynchronization.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1d122ed425f96db46b2098a414911f59
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
