+++ Assets/MixedRealityToolkit.Extensions/SpectatorView/Scripts/StateSynchronization/NetworkExtensions.cs
@@ -8,8 +8,12 @@ using System.Linq;
 using System.Text;
 using UnityEngine;
 
-namespace Microsoft.MixedReality.Toolkit.Extensions.Experimental.Socketer
+namespace Microsoft.MixedReality.Toolkit.Extensions.Experimental.SpectatorView.StateSynchronization
 {
+    /// <summary>
+    /// Extension methods for writing structured CLR and Unity objects to BinaryWriter and reading those
+    /// values from BinaryReader.
+    /// </summary>
     public static class NetworkExtensions
     {
         public static void Write(this BinaryWriter message, Guid value)
