+++ Assets/MixedRealityToolkit.Extensions/SpectatorView/Scripts/StateSynchronization/NetworkExtensions.cs.meta
